Move client() fixture to root conftest.py
