+++ backend/conftest.py
@@ -1,5 +1,22 @@
 import pytest
+from flask import current_app
 from mongoengine import disconnect, connect
+
+from app import app
+
+
+@pytest.fixture(scope='module')
+def client():
+    """Generate test client for every module"""
+    with app.test_client() as client:
+        with app.app_context():
+            app.config.update(
+                TESTING=True,
+                ENV='test'
+            )
+            assert current_app.config["TESTING"]
+            assert current_app.config["ENV"] == "test"
+        yield client
 
 
 @pytest.fixture(scope="function")
